gemspec is now a generated file
While we use syscalls like `git ls-files`.split("\n")` for our convenience,these
kinds of calls combined in a plethora of gems which compounds big/huge Rails
applications sum up to be a bootstrap performance hog.

I've added a rake task called `gemspec`, which uses an ERB template located at
the root of the gem directory to generate the gemspec itself.

More about the bootsrap time
============================

I did some dtracing into the Ruby VM while booting some Rails apps, and it turns
out Bundler actually evaluates the gemspec files of all required gems, and as
fast as `git` is, it's still an external call to an application that has a lot
on its plate (check git internals).

A gemspec generally uses:

`git ls-files`/`Dir.glob("...")` to get the contents of the gem for
files/test_files/executables/spec_files/;
`File.expand_path("...")` to get the contents of signing_key;

In special, the `git` call will chain a **huge mass** of low level stat calls
because of the way it uses the filesystem to store the data.